requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-environ==0.12.0
psycopg2-binary==2.9.10
django-filter==25.1
drf-yasg==1.21.10
faker==37.6.0
